Block hidden link reference closeouts
